+++ app/src/main/res/layout/fragment_group_detail.xml
